requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.64.0

# Academic search
scholarly>=1.7.0
arxiv>=1.4.0

# NLP and text processing
nltk>=3.8
textstat>=0.7.0

# Document processing
python-docx>=0.8.11
markdown2>=2.4.0

# Citation management
pybtex>=0.24.0

# Data handling
pandas>=1.5.0

# Optional: PDF generation
reportlab>=3.6.0

# Development/Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0